Rename table to render
