+++ crates/nu-command/src/default_context.rs
@@ -320,6 +320,7 @@ pub fn add_shell_command_context(mut engine_state: EngineState) -> EngineState {
         // Viewers
         bind_command! {
             Griddle,
+            Render,
             Table,
         };
