+++ crates/nu-command/src/viewers/mod.rs
@@ -1,5 +1,7 @@
 mod griddle;
+mod render;
 mod table;
 
 pub use griddle::Griddle;
+pub use render::Render;
 pub use table::Table;
